+++ src/management/index.ts
@@ -1,3 +1,2 @@
 export * from "./ContextManager";
-export * from "./ContextInterceptor";
 export * from "./ContextWorker";